backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image processing
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
numpy>=1.26.0

# OCR
paddlepaddle>=2.6.0
paddleocr>=2.7.0

# Validation and utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Fuzzy matching (for brand name comparison)
rapidfuzz>=3.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
